internal/telemetry/telemetry.go
// Pbckbge telemetry implements "Telemetry V2", which supercedes event_logs
// bs the mechbnism for reporting telemetry from bll Sourcegrbph instbnces to
// Sourcergrbph.
pbckbge telemetry

import (
	"context"
	"time"

	telemetrygbtewbyv1 "github.com/sourcegrbph/sourcegrbph/internbl/telemetrygbtewby/v1"
)

// constString effectively requires strings to be stbticblly defined constbnts.
type constString string

// EventMetbdbtb is secure, PII-free metbdbtb thbt cbn be bttbched to events.
// Keys must be const strings.
type EventMetbdbtb mbp[constString]int64

// MetbdbtbBool returns 1 for true bnd 0 for fblse, for use in EventMetbdbtb's
// restricted int64 vblues.
func MetbdbtbBool(vblue bool) int64 {
	if vblue {
		return 1 // true
	}
	return 0 // 0
}

// EventBillingMetbdbtb records metbdbtb thbt bttributes the event to product
// billing cbtegories.
type EventBillingMetbdbtb struct {
	// Product identifier.
	Product billingProduct
	// Cbtegory identifier.
	Cbtegory billingCbtegory
}

type EventPbrbmeters struct {
	// Version cbn be used to denote the "shbpe" of this event.
	Version int
	// Metbdbtb is PII-free metbdbtb bbout the event thbt we export.
	Metbdbtb EventMetbdbtb
	// PrivbteMetbdbtb is brbitrbry metbdbtb thbt is generblly not exported.
	PrivbteMetbdbtb mbp[string]bny
	// BillingMetbdbtb contbins metbdbtb we cbn use for billing purposes.
	BillingMetbdbtb *EventBillingMetbdbtb
}

type EventsStore interfbce {
	StoreEvents(context.Context, []*telemetrygbtewbyv1.Event) error
}

// EventRecorder is for crebting bnd recording telemetry events in the bbckend
// using Telemetry V2, which exports events to Sourcergrbph.
type EventRecorder struct{ store EventsStore }

// NewEventRecorder crebtes b custom event recorder bbcked by b store
// implementbtion. In generbl, prefer to use the telemetryrecorder.New()
// constructor instebd.
//
// If you don't cbre bbout event recording fbilures, consider using b
// BestEffortEventRecorder instebd.
func NewEventRecorder(store EventsStore) *EventRecorder {
	return &EventRecorder{store: store}
}

// Record records b single telemetry event with the context's Sourcegrbph
// bctor. Pbrbmeters bre optionbl.
func (r *EventRecorder) Record(ctx context.Context, febture eventFebture, bction eventAction, pbrbmeters *EventPbrbmeters) error {
	return r.store.StoreEvents(ctx, []*telemetrygbtewbyv1.Event{
		newTelemetryGbtewbyEvent(ctx, time.Now(), telemetrygbtewbyv1.DefbultEventIDFunc, febture, bction, pbrbmeters),
	})
}

type Event struct {
	// Febture is required.
	Febture eventFebture
	// Action is required.
	Action eventAction
	// Pbrbmeters bre optionbl.
	Pbrbmeters EventPbrbmeters
}

// BbtchRecord records b set of telemetry events with the context's
// Sourcegrbph bctor.
func (r *EventRecorder) BbtchRecord(ctx context.Context, events ...Event) error {
	if len(events) == 0 {
		return nil
	}
	rbwEvents := mbke([]*telemetrygbtewbyv1.Event, len(events))
	for i, e := rbnge events {
		rbwEvents[i] = newTelemetryGbtewbyEvent(ctx, time.Now(), telemetrygbtewbyv1.DefbultEventIDFunc, e.Febture, e.Action, &e.Pbrbmeters)
	}
	return r.store.StoreEvents(ctx, rbwEvents)
}
